requirements.txt
Flask
pytesseract
APScheduler
Flask-Mail
Werkzeug
Pillow
Flask-SQLAlchemy
python-dotenv
Flask-CORS

